исправления по прослушивателю и комментариям
